FIX the $in serialisation problem
see: https://github.com/devbliss/mongojack/issues/27.

When we do a DBQuery.in("theArray", listString); the plugin was crashing with Execution exception[[ClassCastException: java.lang.String cannot be cast to java.util.Collection]]

I did the modifications for that problem, I check if the value on the SimpleQueryCondition is a String, and if it's a String I just add the value to the serializedConditions.